+++ components/channel/body/ChatCard.tsx
